IntrospectionParser to return Block[] instead of DefsFile.
Change the IntrospectionParser method to parse data to return a map
of Block arrays instead of a map of DefsFile. This allow to go back
to the previous behaviour of a DefsFile (which is not adding or
changing Block) and to override or add a Block before creating the
DefsFile instance.