Use UnnecessaryCodeException instead of IgnoreIntrospectionException.
IgnoreIntrospectionException was an Exception subclass made to help
the IntrospectionParser knows when it will have to skip the parsing
of an element. UnnecessaryCodeException was made for this purpose
but for the DefsParser instead. So it seens legit to use this already
existing Exception subclass to handle the cases to ignore with the
IntrospectionParser.