Introduce the IntrospectionRepository object.
Add the IntrospectionRepository class which represents the
<repository> element of GIR data. This object has references to the
<namespace> elements that it contains, references to the headers that
need to be included in the generated code and it also has knowledge
of all the c:identifier-prefixes attributes values found in all the
<namespace> elements found.
The IntrospectionParser now takes IntrospectionRepository as input
instead of Reader objects.
This change is made to remove the hard code list of prefixes used
before. For this we need to load every XML files into memory and
grab the values of c:identifier-prefixes attributes. The <namespace>
elements are then parsed by the IntrospectionParser as it used to be
before.