Change types.list format to use XML.
Use XML format for the types.list file. It must contains a <list>
element which will have <type> childs. Each <type> has a name that
is the C name of a type and contains <header> childs that give
headers files to include in the generated C code and <thing> childs
that tell which function are to be ignored by the IntrospectionParser
using the 'ignore' attribute.
The idea behind using the XML format is to take advantage of our
knowledge on parsing XML with a well working library and to extend
the possibilty of what can be done with the types.list file.