Java class name override in types.list.
Use the new types.list file to provide a specific Java name for a
given C type. This is done by using the 'java-name' attribute of a
<type> element. The parser will use the given Java name as name for
the public API.
This change allows the removal of the names-override.properties file
and all the ode associated to its parsing and use.