Java package name overridein types.list.
Use the new types.list file to provide a specific Java package name
for a given module/GIR namespace. This is done by using the <module>
element with the 'name' attribute being the module/GIR namespace and
the 'java-package' attribute being the name to use in the Java
context. This Java package name must at least start with an upper
case letter because of its use in the Java class name in the
translation layer.
This change allows the removal of the packages-override.properties
file and all the code associated to its parsing and use.